@@ -1,12 +1,11 @@
-namespace CameraWatch.Models
+namespace CameraWatch.Models;
+
+public class AppConfig
 {
-    public class AppConfig
-    {
-        public string LogFile1Path { get; set; } = string.Empty;
-        public string LogFile2Path { get; set; } = string.Empty;
-        public int DisplayDurationSeconds { get; set; } = 10;
-        public int DisplayAreaWidth { get; set; } = 576;
-        public int DisplayAreaHeight { get; set; } = 192;
-    }
+    public string LogFile1Path { get; set; } = string.Empty;
+    public string LogFile2Path { get; set; } = string.Empty;
+    public int DisplayDurationSeconds { get; set; } = 10;
+    public int DisplayAreaWidth { get; set; } = 576;
+    public int DisplayAreaHeight { get; set; } = 192;
 }
 
